add delete edit list adminProduct

+++ admin/modules/category/list.php
@@ -6,8 +6,11 @@ $data = [
 layout('header','admin', $data);
 layout('sidebar','admin', $data);
 layout('breadcrumb','admin', $data);
+$msg = getFlashData('msg');
+$msgType = getFlashData('msg_type');
 ?>
 <!-- Main content -->
+<?php getMsg($msg, $msgType); ?>
 <section class="content">
     <div class="container-fluid">
         <table class="table table-bordered">
@@ -23,9 +26,9 @@ layout('breadcrumb','admin', $data);
                 <tbody>
                 <?php
                 $i= 1;
-                $query = getRaw("SELECT * FROM categories");
-                if (is_array($query)):
-                foreach($query as $value):
+                $categoryQuery = getRaw("SELECT * FROM categories");
+                if (is_array($categoryQuery)):
+                foreach($categoryQuery as $value):
                 ?>
                 <tr>
                     <td><?php echo $i++; ?></td>
@@ -33,7 +36,7 @@ layout('breadcrumb','admin', $data);
                     <td><?php echo $value['created_at'];?></td>
                     <td><?php echo $value['updated_at'];?></td>
                     <td><a href="?module=category&action=edit&id=<?php echo $value['id'];?>" class="btn btn-primary">Sửa</a></td>
-                    <td><a href="?module=category&action=delete&id=<?php echo $value['id'];?>" class="btn btn-danger">Xóa</a></td>
+                    <td><a href="?module=category&action=delete&id=<?php echo $value['id'];?>" onclick="return confirm('Bạn chắc chắn muốn xóa chứ?');" class="btn btn-danger">Xóa</a></td>
                 </tr>
                 <?php endforeach; endif;?>
                 </tbody>
